[INJIVER-1065] refactor verify another qr code (#391)
* [INJIVER846] removed unused method


* [INJIVER846] refactor cliend Id validation


* [INJIVER-846] refactor config json


* [INJIVER-846] refactor configmaps on deployment


* [INJIVER-846] refactor configmaps on deployment


* [INJIVER-846] refactor configmaps on deployment


* [INJIVER-846] refactor configmaps on deployment


* [INJIVER-846] refactor configmaps on deployment


* [INJIVER-846] refactor configmaps on deployment


* [INJIVER-846] refactor configmaps on deployment


* refactor rerendering order and default order


* [INJIVER-1024] version upgrade to verify mock vc


* [INJIVER-1024] vc-verifier version downgrade


* [INJIVER-1024] vc-verifier version downgrade


* [INJIVER-846] fix nginx config to send large data and upgarding vc-verifier library


* [INJIVER-456] fix error message alignment


* [INJIVER-846] refactor default rendering order


* [INJIVER-1066] fix missing element ID


* [INJIVER-1066] refactor push trigger workflow


* [INJIVER-1065] refactor verify another qr code


---------
