+++ verify-ui/src/components/Home/VerificationSection/Result/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import ResultSummary from "./ResultSummary";
 import { useVerificationFlowSelector } from "../../../../redux/features/verification/verification.selector";
 import DisplayVcDetailsModal from "./DisplayVcDetailsModal";
@@ -6,7 +6,7 @@ import DisplayVcDetailView from "./DisplayVcDetailView";
 import { Button } from "../commons/Button";
 import { useTranslation } from "react-i18next";
 import { useAppDispatch } from "../../../../redux/hooks";
-import {goToHomeScreen} from "../../../../redux/features/verification/verification.slice";
+import { acceptedFileTypes, handleFileUpload } from "../../../../utils/fileUploadUtils";
 
 const Result = () => {
   const { vc, vcStatus } = useVerificationFlowSelector((state) => state.verificationResult ?? { vc: null, vcStatus: null });
@@ -14,7 +14,9 @@ const Result = () => {
   const credentialType: string = vc.type[1];
   const {t} = useTranslation();
   const dispatch = useAppDispatch();
-
+  const handleUploadQrCode = () => {
+    document.getElementById("verify-another-qrcode")?.click();
+  };
   // validate vc and show success/failure component
   return (
     <div id="result-section" className="relative mb-[100px]">
@@ -32,8 +34,16 @@ const Result = () => {
           <Button
             id="verify-another-qr-code-button"
             title={t("Common:Button.verifyAnotherQrCode")}
-            onClick={()=>dispatch(goToHomeScreen({}))}
+            onClick={handleUploadQrCode}
             className="mx-auto mt-6 mb-20 lg:mb-6 lg:w-[339px]"
+          />
+          <input
+            type="file"
+            id="verify-another-qrcode"
+            name="upload-qr"
+            accept={acceptedFileTypes}
+            className="mx-auto my-2 hidden h-0"
+            onChange={(e) => handleFileUpload(e, dispatch)}
           />
         </div>
       </div>
