+++ verify-ui/src/components/Home/VerificationSection/UploadQrCode.tsx
@@ -1,65 +1,9 @@
-import { scanFilesForQr } from "../../../utils/qr-utils";
-import {
-  AlertMessages,
-  SupportedFileTypes,
-  UploadFileSizeLimits,
-} from "../../../utils/config";
 import { GradientUploadIcon, WhiteUploadIcon } from "../../../utils/theme-utils";
 import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
-import {
-  goToHomeScreen,
-  qrReadInit,
-  verificationInit,
-} from "../../../redux/features/verification/verification.slice";
-import { raiseAlert } from "../../../redux/features/alerts/alerts.slice";
-import { checkInternetStatus, getFileExtension } from "../../../utils/misc";
-import { updateInternetConnectionStatus } from "../../../redux/features/application-state/application-state.slice";
-import { AlertInfo } from "../../../types/data-types";
-import { Dispatch } from "redux";
 import { useState } from "react";
 import { RootState } from "../../../redux/store";
 import { isRTL } from "../../../utils/i18n";
-
-const doFileChecks = (dispatch: Dispatch, file: File | null): boolean => {
-  if (!file) return false;
-  let alert: AlertInfo | null = null;
-  // file format check
-  const fileExtension = getFileExtension(file.name).toLowerCase();
-  if (!SupportedFileTypes.includes(fileExtension)) {
-    alert = AlertMessages().unsupportedFileType;
-  }
-
-  // file size check
-  if (
-    file.size < UploadFileSizeLimits.min ||
-    file.size > UploadFileSizeLimits.max
-  ) {
-    alert = AlertMessages().unsupportedFileSize;
-  }
-
-  if (alert) {
-    dispatch(goToHomeScreen({}));
-    dispatch(raiseAlert({ ...alert, open: true }));
-    return false;
-  }
-  return true;
-};
-
-const doInternetCheck = async (dispatch: Dispatch) => {
-  dispatch(
-    updateInternetConnectionStatus({ internetConnectionStatus: "LOADING" })
-  );
-  let isOnline = await checkInternetStatus();
-  dispatch(
-    updateInternetConnectionStatus({
-      internetConnectionStatus: isOnline ? "ONLINE" : "OFFLINE",
-    })
-  );
-  return isOnline;
-};
-
-const acceptedFileTypes = SupportedFileTypes.map((ext) => `.${ext}`).join(", ");
-
+import { acceptedFileTypes, handleFileUpload } from "../../../utils/fileUploadUtils";
 
 export const UploadQrCode = ({
   displayMessage,
@@ -96,10 +40,8 @@ export const UploadQrCode = ({
   }
 
   return (
-    <div
-      className={`mx-auto my-1.5 flex content-center justify-center ${className}`}
-    >
-      <UploadButton displayMessage={displayMessage} />
+    <div className={`mx-auto my-1.5 flex content-center justify-center ${className}`} >
+      <UploadButton displayMessage={displayMessage} />     
       <br />
       <input
         type="file"
@@ -107,37 +49,7 @@ export const UploadQrCode = ({
         name="upload-qr"
         accept={acceptedFileTypes}
         className="mx-auto my-2 hidden h-0"
-        onChange={async (e) => {
-          const isOnline = await doInternetCheck(dispatch);
-          if (!isOnline) return;
-
-          const file = e?.target?.files && e?.target?.files[0];
-          const fileChecksPassed = doFileChecks(dispatch, file);
-          if (!fileChecksPassed) {
-            if (e?.target) e.target.value = ""; // clear the target to be able to read same file again
-            return;
-          }
-
-          scanFilesForQr(file).then((scanResult) => {
-            if (scanResult.error) console.error(scanResult.error);
-            if (!!scanResult.data) {
-              dispatch(qrReadInit({ method: "UPLOAD" }));
-              dispatch(
-                raiseAlert({ ...AlertMessages().qrUploadSuccess, open: true })
-              );
-              dispatch(
-                verificationInit({
-                  qrReadResult: { qrData: scanResult.data, status: "SUCCESS" },
-                })
-              );
-            } else {
-              dispatch(
-                raiseAlert({ ...AlertMessages().qrNotDetected, open: true })
-              );
-              dispatch(goToHomeScreen({}));
-            }
-          });
-        }}
+        onChange={(e) => handleFileUpload(e, dispatch)}
       />
     </div>
   );
